+++ package.json
@@ -113,7 +113,7 @@
     "url-search-params-polyfill": "8.1.1",
     "vite": "4.3.5",
     "vite-plugin-environment": "^1.1.3",
-    "vue-tsc": "1.6.5",
+    "vue-tsc": "1.8.8",
     "vue3-gettext": "^2.3.3",
     "wait-for-expect": "3.0.2"
   },
